backend/pyproject.toml
[project]
name = "fluencyai-backend"
version = "0.1.0"
description = "FluencyAI FastAPI backend"
requires-python = "==3.12.*"
dependencies = [
  "alembic==1.14.0",
  "bcrypt==4.2.1",
  "email-validator==2.2.0",
  "fastapi==0.115.6",
  "httpx==0.28.1",
  "psycopg==3.2.3",
  "psycopg-binary==3.2.3",
  "pydantic-settings==2.7.1",
  "PyJWT==2.10.1",
  "SQLAlchemy==2.0.36",
  "uvicorn==0.34.0",
  "google-generativeai==0.8.6",
  "pypdf==6.10.2",
]

[project.optional-dependencies]
dev = [
  "pytest==8.3.4",
  "pytest-cov==6.0.0",
]

[build-system]
requires = ["setuptools>=80.0.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = [
  "application*",
  "domain*",
  "infrastructure*",
  "presentation*",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
addopts = "-q"

[tool.coverage.run]
source = ["domain", "application", "infrastructure", "presentation"]

[tool.coverage.report]
fail_under = 80
show_missing = true
